refactor(agent-gui): split detail pane props
