+++ packages/agent/gui/agent-gui/agentGuiNode/view/AgentGUIDetailPane.tsx
@@ -1,36 +1,12 @@
 import { memo, useCallback, useEffect, useMemo, useRef, useState } from "react";
-import type { WorkspaceUserProjectI18nRuntime } from "@tutti-os/workspace-user-project/i18n";
-import type { WorkspaceLinkAction } from "../../../actions/workspaceLinkActions";
-import type { UiLanguage } from "../../../contexts/settings/domain/agentSettings";
 import type { AgentMessageMarkdownWorkspaceAppIcon } from "../../../shared/AgentMessageMarkdown";
 import { latestAssistantMessageText } from "../../../shared/agentConversation/projection/agentConversationProjection";
 import { AGENT_GUI_WORKBENCH_OPEN_EXTERNAL_IMPORT_EVENT } from "../../../workbench/contribution";
 import { resolveAgentGuiWorkbenchProviderLabel } from "../../../workbench/providerCatalog";
-import type {
-  AgentComposerGitBranchLoader,
-  AgentComposerProps,
-  AgentComposerSlashStatusLimit,
-  WorkspaceReferencePickResult
-} from "../AgentComposer";
-import type { AgentContextMentionItem } from "../agentRichText/agentFileMentionExtension";
-import type {
-  AgentGUIComposerViewModel,
-  AgentGUIDetailViewModel,
-  AgentHomeSuggestionAction,
-  AgentGUIInteractionViewModel,
-  AgentGUIOperationsViewModel,
-  AgentGUIRailViewModel,
-  AgentGUIReadinessViewModel,
-  AgentGUIShellViewModel
-} from "../model/agentGuiNodeTypes";
+import type { AgentComposerProps } from "../AgentComposer";
+import type { AgentHomeSuggestionAction } from "../model/agentGuiNodeTypes";
 import { updateAgentComposerDraft } from "../model/agentComposerDraft";
 import { resolveAgentComposerDraftScopeKey } from "../model/agentComposerDraftScope";
-import type { AgentGUIManagedHomeTargetProjection } from "../model/agentGuiProviderRailOrder";
-import type {
-  AgentGUINodeViewProps,
-  AgentGUIProviderUnavailableStateRenderer,
-  AgentGUIViewLabels
-} from "../AgentGUINodeView";
 import {
   buildAgentConversationHandoffPrompt,
   handoffProjectPathForConversation
@@ -51,58 +27,11 @@ import styles from "../AgentGUINode.styles";
 import { useAgentGUIDetailScroll } from "./useAgentGUIDetailScroll";
 import { useAgentGUIDetailModel } from "./useAgentGUIDetailModel";
 import { useAgentGUIComposerInputHistoryProps } from "./useAgentGUIComposerInputHistoryProps";
-import type { AgentGUIComposerEngagement } from "../engagement/agentGUIEngagement.types";
 import { useAgentGUITuttiWorkflow } from "./useAgentGUITuttiWorkflow";
 import type { AgentTranscriptVirtualScrollController } from "../../../shared/agentConversation/components/AgentTranscriptView";
+import type { AgentGUIDetailPaneProps } from "./AgentGUINodeView.types";
 export const EMPTY_WORKSPACE_APP_ICONS: readonly AgentMessageMarkdownWorkspaceAppIcon[] =
   [];
-export interface AgentGUIDetailPaneProps {
-  shell: AgentGUIShellViewModel;
-  rail: AgentGUIRailViewModel;
-  detail: AgentGUIDetailViewModel;
-  composer: AgentGUIComposerViewModel;
-  interaction: AgentGUIInteractionViewModel;
-  readiness: AgentGUIReadinessViewModel;
-  operations: AgentGUIOperationsViewModel;
-  homeTargetProjection: AgentGUIManagedHomeTargetProjection;
-  referenceProvenanceFilters?: AgentComposerProps["referenceProvenanceFilters"];
-  sessionInputHistoryEnabled?: boolean;
-  composerEngagement?: AgentGUIComposerEngagement;
-  actions: AgentGUINodeViewProps["actions"];
-  labels: AgentGUIViewLabels;
-  workspaceUserProjectI18n: WorkspaceUserProjectI18nRuntime;
-  uiLanguage: UiLanguage;
-  isActive: boolean;
-  isVisible: boolean;
-  workspaceReferencePickerOpen: boolean;
-  composerFocusRequestSequence: number | null;
-  slashStatusLimits: readonly AgentComposerSlashStatusLimit[];
-  slashStatusLimitsLoading: boolean;
-  slashStatusLimitsUnavailable: boolean;
-  slashStatusOverride?: AgentComposerProps["slashStatus"];
-  onSlashStatusOpen?: AgentComposerProps["onSlashStatusOpen"];
-  onSlashStatusClose?: AgentComposerProps["onSlashStatusClose"];
-  onSlashStatusRefresh?: AgentComposerProps["onSlashStatusRefresh"];
-  onLinkAction?: (action: WorkspaceLinkAction) => void;
-  onHandoffConversation?: AgentGUINodeViewProps["onHandoffConversation"];
-  capabilityMenuState?: AgentComposerProps["capabilityMenuState"];
-  capabilityControlsReadOnly?: AgentComposerProps["capabilityControlsReadOnly"];
-  onCapabilitySettingsRequest?: AgentComposerProps["onCapabilitySettingsRequest"];
-  onAgentProviderLogin?: (provider?: string | null) => void;
-  onRequestWorkspaceReferences?:
-    | ((
-        entity?: AgentContextMentionItem | null
-      ) => Promise<WorkspaceReferencePickResult>)
-    | null;
-  resolveExternalPromptEntries?: AgentComposerProps["resolveExternalPromptEntries"];
-  prepareExternalPromptFiles?: AgentComposerProps["prepareExternalPromptFiles"];
-  promptAssetLimit?: number | null;
-  selectProjectDirectory?: () => Promise<{ path: string } | null>;
-  onRequestGitBranches?: AgentComposerGitBranchLoader | null;
-  onRequestComposerFocus: () => void;
-  workspaceAppIcons?: readonly AgentMessageMarkdownWorkspaceAppIcon[];
-  renderProviderUnavailableState?: AgentGUIProviderUnavailableStateRenderer;
-}
 export const AgentGUIDetailPane = memo(function AgentGUIDetailPane({
   shell,
   rail,
